pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marketing-adk-agent"
version = "0.1.0"
description = "AI Marketing Copywriter Agent with Gradio UI"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.0.0",
    "gradio>=6.1.0",
    "python-dotenv>=1.0.0",
]

